Create IOUtilities class
