+++ AudioTagger.Library/MediaFile.cs
@@ -180,7 +180,7 @@ public class MediaFile
                                 searchSubDirectories
                                     ? System.IO.SearchOption.AllDirectories
                                     : System.IO.SearchOption.TopDirectoryOnly)
-                .Where(FileSelection.Filter)
+                .Where(IOUtilities.IsSupportedFileExtension)
                 .ToArray();
 
             foreach (var fileName in fileNames)
@@ -197,7 +197,7 @@ public class MediaFile
                                 .ToList();
         }
 
-        // If the path is a file
+        // Otherwise, if the path is a file
         if (System.IO.File.Exists(path))
         {
             return new List<MediaFile> { MediaFileFactory.CreateFileData(path) };
@@ -246,37 +246,5 @@ public class MediaFile
             return false;
 
         return true;
-    }
-
-    /// <summary>
-    /// Replaces characters that are invalid in file path names with a specified safe character.
-    /// </summary>
-    /// <returns>A corrected string or the original if no changes were needed.</returns>
-    /// <remarks>TODO: Make a new class for this (e.g., FileUtilities, etc.).</remarks>
-    public static string EnsurePathSafeString(string input, char replacementChar = '_')
-    {
-        return System.IO.Path.GetInvalidFileNameChars()
-                    .ToList()
-                    .Aggregate(
-                        new StringBuilder(input),
-                        (workingString, invalidChar) =>
-                            workingString.Replace(invalidChar, replacementChar))
-                    .ToString();
-    }
-
-    /// <summary>
-    /// Concatenates multiple inputs to a string, then replaces characters that are
-    /// invalid in file path names with a specified safe character.
-    /// </summary>
-    /// <returns>A corrected string or the original if no changes were needed.</returns>
-    /// <remarks>TODO: Make a new class for this (e.g., FileUtilities, etc.).</remarks>
-    public static string EnsurePathSafeString(
-        IEnumerable<string> input,
-        char replacementChar = '_',
-        string joinWith = " && ")
-    {
-        return EnsurePathSafeString(
-            string.Join(joinWith, input),
-            replacementChar);
     }
 }
